pyproject.toml
[project]
name = "doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.121.2",
    "langchain-community>=0.3.27",
    "pypdf>=6.2.0",
    "pypdf2>=3.0.1",
]
